projects/sequence-web/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
httpx>=0.27.0
MLB-StatsAPI>=1.7.2
python-dotenv>=1.0.0
thefuzz>=0.22.0
pandas>=2.0.0
matplotlib>=3.7.0
aiofiles>=23.2.1
